% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Load and clean the NOAA Significant equation dataset}
\usage{
eq_clean_data(fileLink, fileName = "signif.txt")
}
\arguments{
\item{fileLink}{string or connection. A path to the NOAA dataset.}

\item{fileName}{string. The dataset name (if renamed).}
}
\description{
\code{eq_clean_data} loads the NOAA dataset into memory and cleans it for easier mapping.
The user provides a path to the data set and its name.
}
\examples{
\donttest{
eq_clean_data('../')
}

}
