% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create labels for earthquake maps}
\usage{
eq_create_label(loc, magnitude, deaths)
}
\arguments{
\item{loc}{A character vector that contains locations.}

\item{magnitude}{A numeric vector that contains earthquake magnitudes.}

\item{deaths}{A numeric vector that contains death counts.}
}
\value{
An HTML-formatted string (strings) that is used as a label
}
\description{
\code{eq_create_label} creates labels for earthquake maps.
The labels include: location, magnitude, and the total number of deaths.
}
\examples{
eq_create_label('Mexico', 8.0, 2000)

}
