% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_data}
\alias{eq_data}
\title{Parameters for 1000 earthquake events}
\format{A data frame with 1000 rows and 46 variables.}
\usage{
eq_data
}
\description{
A subset of the NOAA Significant Earthquakes dataset.
}
\keyword{datasets}
