% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location data from the NOAA data set}
\usage{
eq_location_clean(locCol)
}
\arguments{
\item{locCol}{A character vector. Location strings that need to be converted.}
}
\description{
\code{eq_location_clean} extracts a first string preceded by a colon in the LOCATION column
and converts it to the title case.
}
\examples{
eq_location_clean('COLOMBIA: BOGOTA')
\donttest{
data("eq_data"); eq_location_clean(eq_data$LOCATION_NAME)
}

}
