% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{Create earthquake maps}
\usage{
eq_map(data = data, annot_col = "DATE")
}
\arguments{
\item{data}{A data frame (tibble). A subset of the NOAA Significant Earthquakes dataset. This dataset is used
for mapping.}

\item{annot_col}{A string. A dataset column from the dataset used for earthquake labels.
This column can be created using \code{eq_create_label}.}
}
\description{
\code{eq_map} creates a regional earthquake map using a subset of the NOAA Significant Earthquake dataset.
}
\examples{
\donttest{data("eq_data"); eq_data \%>\% filter(COUNTRY == 'MEXICO') \%>\%
eq_map()
}
\donttest{data("eq_data"); eq_data \%>\% filter(COUNTRY == 'MEXICO') \%>\%
eq_map(annot_col = 'EQ_PRIMARY')
}

}
