% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Timeline plots}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeline
}
\description{
A geom object used with \code{geom_timeline}.
}
\keyword{datasets}
