% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{stat_timeline}
\alias{stat_timeline}
\title{Timeline plots}
\usage{
stat_timeline(mapping = NULL, data = NULL, geom = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
A stat function that pre-processes data for \code{geom_timeline} and
 \code{geom_timeline_label}.
}
