% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Timeline plots}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimeline
}
\description{
A stat object that used for data preprocessing in \code{geom_timeline} and
  \code{geom_timeline_label}.
}
\keyword{datasets}
