% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{theme_eq}
\alias{theme_eq}
\title{A custom theme for timeline plots}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 59.}
\usage{
theme_eq
}
\description{
A custom theme for timeline plots
}
\examples{
\donttest{
data("eq_data")
ggplot(data = eq_data \%>\% filter(COUNTRY \%in\% c('COLOMBIA', 'MEXICO', 'USA')),
       aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY,
           color = TOTAL_DEATHS, xmin = as.Date('1970-01-01'),
           xmax = as.Date('2015-01-01'))) +
  geom_timeline_label(aes(label = as.character(DATE)), nmax = 2) +
  geom_timeline() +
  theme_eq
}

}
\keyword{datasets}
