% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline plots}
\usage{
geom_timeline(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{xmin}{a date. The lower bound of the date interval used for subsetting the NOAA data.}

\item{xmax}{a date. The upper bound of the date interval used for subsetting the NOAA data.}

\item{nmax}{an integer. Top earthquakes with respect to the magnitude. These earthquakes are
labelled on the plot.}

\item{label}{a character vector. A column in the NOAA dataset used for labelling the data points.}
}
\description{
\code{geom_timeline} is used to visualize earthquake occurence, intensity,
  the number of deaths caused, and other parameters.
}
\examples{
\donttest{
data("eq_data")
ggplot(data = eq_data \%>\% filter(COUNTRY == "ITALY"),
       aes(x = DATE, size = EQ_PRIMARY, color = TOTAL_DEATHS,
           xmin = as.Date('1950-01-01'),
           xmax = as.Date('2015-01-01'))) +
 geom_timeline() +
 theme_eq
}
\donttest{
data("eq_data")
ggplot(data = eq_data \%>\% filter(COUNTRY \%in\% c('COLOMBIA', 'MEXICO', 'USA')),
       aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY,
           color = TOTAL_DEATHS, xmin = as.Date('1970-01-01'),
           xmax = as.Date('2015-01-01'))) +
  geom_timeline_label(aes(label = as.character(DATE)), nmax = 2) +
  geom_timeline() +
  theme_eq
}
}
