% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline plots (labels)}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{nmax}{an integer. Top earthquakes with respect to the magnitude. These earthquakes are
labelled on the plot.}

\item{label}{a character vector. A column in the NOAA dataset used for labelling the data points.}
}
\description{
\code{geom_timeline_label} generates labels for earthquakes on a timeline plot. Used along with \code{geom_timeline}.
}
\examples{
\donttest{
data("eq_data")
ggplot(data = eq_data \%>\% filter(COUNTRY \%in\% c('COLOMBIA', 'MEXICO', 'USA')),
       aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY,
           color = TOTAL_DEATHS, xmin = as.Date('1970-01-01'),
           xmax = as.Date('2015-01-01'))) +
  geom_timeline_label(aes(label = as.character(DATE)), nmax = 2) +
  geom_timeline() +
  theme_eq
}
}
