% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{theme_eq}
\alias{theme_eq}
\title{A custom theme for timeline plots}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 59.}
\usage{
theme_eq
}
\description{
A custom theme for timeline plots
}
\keyword{datasets}
